+++ CHANGELOG.md
@@ -11,6 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+* Installers: Added OS name to the installer filenames
+* Improved Update Download dialog. Now shows the percentage and downloaded size
+
 ### Deprecated
 
 ### Removed